requirements.txt
flask
flask-bcrypt
flask-wtf
flask-talisman
mysql-connector-python
gunicorn